perf: look up equals-only groups directly and shrink value references
Most groups only compare with equals. For them, evalGroup now looks the
values up in the hash table itself instead of going through collect.
Value references use 32-bit offsets, which keeps them at 32 bytes
despite the cached number.

+++ matchjson.go
@@ -115,7 +115,7 @@ func (sc *scratch) groupJSONValues() {
 	sc.frefs = sc.frefs[:n]
 	for _, v := range sc.jvals {
 		sp := &sc.spans[v.span]
-		sc.frefs[sp.fhi] = vref{lo: v.lo, hi: v.hi}
+		sc.frefs[sp.fhi] = vref{lo: uint32(v.lo), hi: uint32(v.hi)}
 		sp.fhi++
 	}
 }
